ステート関連ファイルをコンポーザブルに移動
