+++ src/composables/sequencerStateMachine/common.ts
