+++ src/composables/sequencerStateMachine/states/idleState.ts
@@ -5,7 +5,7 @@ import {
   getGuideLineTicks,
   Input,
   SequencerStateDefinitions,
-} from "@/components/Sing/sequencerStateMachine/common";
+} from "@/composables/sequencerStateMachine/common";
 import { getButton, isSelfEventTarget } from "@/sing/viewHelper";
 import { isOnCommandOrCtrlKeyDown } from "@/store/utility";
