+++ src/composables/sequencerStateMachine/states/moveNoteState.ts
@@ -1,5 +1,6 @@
 import { getOrThrow } from "@/helpers/mapHelper";
 import { State, SetNextState } from "@/sing/stateMachine";
+import { clamp } from "@/sing/utility";
 import { getButton, PREVIEW_SOUND_DURATION } from "@/sing/viewHelper";
 import { Note } from "@/store/type";
 import { TrackId, NoteId } from "@/type/preload";
@@ -9,13 +10,12 @@ import {
   Input,
   PositionOnSequencer,
   SequencerStateDefinitions,
-} from "@/components/Sing/sequencerStateMachine/common";
-import { clamp } from "@/sing/utility";
+} from "@/composables/sequencerStateMachine/common";
 
-export class ResizeNoteLeftState
+export class MoveNoteState
   implements State<SequencerStateDefinitions, Input, Context>
 {
-  readonly id = "resizeNoteLeft";
+  readonly id = "moveNote";
 
   private readonly cursorPosAtStart: PositionOnSequencer;
   private readonly targetTrackId: TrackId;
@@ -51,7 +51,7 @@ export class ResizeNoteLeftState
     this.currentCursorPos = args.cursorPosAtStart;
   }
 
-  private previewResizeLeft(context: Context) {
+  private previewMove(context: Context) {
     if (this.innerContext == undefined) {
       throw new Error("innerContext is undefined.");
     }
@@ -64,20 +64,24 @@ export class ResizeNoteLeftState
     const newNotePos =
       Math.round((notePos + dragTicks) / snapTicks) * snapTicks;
     const movingTicks = newNotePos - notePos;
+    const movingSemitones =
+      this.currentCursorPos.noteNumber - this.cursorPosAtStart.noteNumber;
 
     const editedNotes = new Map<NoteId, Note>();
     for (const note of previewNotes) {
       const targetNoteAtStart = getOrThrow(targetNotesAtStart, note.id);
-      const notePos = targetNoteAtStart.position;
-      const noteEndPos =
-        targetNoteAtStart.position + targetNoteAtStart.duration;
-      const position = clamp(notePos + movingTicks, 0, noteEndPos - snapTicks);
-      const duration = noteEndPos - position;
+      const position = Math.max(0, targetNoteAtStart.position + movingTicks);
+      const noteNumber = clamp(
+        targetNoteAtStart.noteNumber + movingSemitones,
+        0,
+        127,
+      );
 
-      if (note.position !== position || note.duration !== duration) {
-        editedNotes.set(note.id, { ...note, position, duration });
+      if (note.position !== position || note.noteNumber !== noteNumber) {
+        editedNotes.set(note.id, { ...note, noteNumber, position });
       }
     }
+
     if (editedNotes.size !== 0) {
       context.previewNotes.value = previewNotes.map((value) => {
         return editedNotes.get(value.id) ?? value;
@@ -85,7 +89,8 @@ export class ResizeNoteLeftState
       this.innerContext.edited = true;
     }
 
-    context.guideLineTicks.value = newNotePos;
+    context.guideLineTicks.value =
+      this.innerContext.guideLineTicksAtStart + movingTicks;
   }
 
   onEnter(context: Context) {
@@ -106,7 +111,7 @@ export class ResizeNoteLeftState
         throw new Error("innerContext is undefined.");
       }
       if (this.innerContext.executePreviewProcess) {
-        this.previewResizeLeft(context);
+        this.previewMove(context);
         this.innerContext.executePreviewProcess = false;
       }
       this.innerContext.previewRequestId =
@@ -139,11 +144,10 @@ export class ResizeNoteLeftState
       if (input.mouseEvent.type === "mousemove") {
         this.currentCursorPos = input.cursorPos;
         this.innerContext.executePreviewProcess = true;
-      } else if (
-        input.mouseEvent.type === "mouseup" &&
-        mouseButton === "LEFT_BUTTON"
-      ) {
-        setNextState("idle", undefined);
+      } else if (input.mouseEvent.type === "mouseup") {
+        if (mouseButton === "LEFT_BUTTON") {
+          setNextState("idle", undefined);
+        }
       }
     }
   }
@@ -161,7 +165,9 @@ export class ResizeNoteLeftState
       notes: previewNotes,
       trackId: this.targetTrackId,
     });
-    void context.store.actions.SELECT_NOTES({ noteIds: previewNoteIds });
+    void context.store.actions.SELECT_NOTES({
+      noteIds: previewNoteIds,
+    });
 
     if (previewNotes.length === 1) {
       void context.store.actions.PLAY_PREVIEW_SOUND({
